requirements.txt
streamlit==1.35.0
python-dotenv==1.0.0
pandas==2.1.4
tabulate==0.9.0
